type conversion

+++ src/models/layers/decoder.py
@@ -185,7 +185,7 @@ class BaseRNNDecoder(nn.Module):
             #       Each element is beam index: 0 ~ beam_size
             #                     + position index: 0 ~ beam_size x (batch_size-1)
             beam_idx = top_k_idx / self.vocab_size  # [batch_size, beam_size]
-            top_k_pointer = (beam_idx + batch_position.unsqueeze(1)).view(-1)
+            top_k_pointer = (beam_idx + batch_position.unsqueeze(1)).view(-1).long()
 
             # Select next h (size doesn't change)
             # [num_layers, batch_size * beam_size, hidden_size]
